Geração de Conteúdo Inteligente/viral-content-brain/backend/requirements.txt
# Core
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic-settings==2.5.2
python-dotenv==1.0.1

# DB
sqlalchemy[asyncio]==2.0.35
asyncpg==0.29.0
alembic==1.13.3
pgvector==0.3.3

# Ingestion
apify-client==1.8.1
httpx==0.27.2
tenacity==9.0.0

# Media processing
ffmpeg-python==0.2.0
Pillow==10.4.0

# LLM / AI
anthropic==0.36.0
openai==1.51.0

# Worker
arq==0.26.1
redis==5.1.1

# CLI
typer==0.12.5
rich==13.9.2

# Utils
python-slugify==8.0.4
